[MIG] stock_auto_move to v9.0
