+++ stock_auto_move/data/stock_auto_move_data.xml
@@ -1,11 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
-<openerp>
-    <data>
 
-        <record model="procurement.group" id="automatic_group">
-            <field name="name">Automatic</field>
-            <field name="move_type">direct</field>
-        </record>
+<odoo>
 
-    </data>
-</openerp>
+    <record model="procurement.group" id="automatic_group">
+        <field name="name">Automatic</field>
+        <field name="move_type">direct</field>
+    </record>
+
+</odoo>
