+++ stock_auto_move/models/__init__.py
@@ -1,2 +1,5 @@
 # -*- coding: utf-8 -*-
-from . import stock_auto_move
+from . import stock_move
+from . import procurement_rule
+from . import procurement_order
+from . import stock_location_path
