+++ stock_auto_move/models/procurement_order.py
@@ -1,0 +1,16 @@
+# -*- coding: utf-8 -*-
+# © 2014-2015 NDP Systèmes (<http://www.ndp-systemes.fr>)
+
+from openerp import api, models
+
+
+class ProcurementOrder(models.Model):
+
+    _inherit = 'procurement.order'
+
+    @api.model
+    def _run_move_create(self, procurement):
+        res = super(ProcurementOrder, self)._run_move_create(procurement)
+        if procurement.rule_id:
+            res.update({'auto_move': procurement.rule_id.auto_move})
+        return res
