+++ stock_auto_move/views/procurement_rule.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!-- Copyright 2017 ACSONE SA/NV
+     License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl). -->
+
+<odoo>
+
+    <record model="ir.ui.view" id="stock_auto_move_procurement_rule_form_view">
+        <field name="name">stock.auto.move.procurement.rule.form.view</field>
+        <field name="model">procurement.rule</field>
+        <field name="inherit_id" ref="stock.view_procurement_rule_form_stock_inherit"/>
+        <field name="priority" eval="16"/>
+        <field name="arch" type="xml">
+            <field name="procure_method" position="after">
+                <field name="auto_move"/>
+            </field>
+        </field>
+    </record>
+
+</odoo>
