--- stock_auto_move/views/stock_auto_move_view.xml
@@ -1,49 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<openerp>
-    <data>
-
-        <record model="ir.ui.view" id="stock_auto_move_move_form_view">
-            <field name="name">stock.auto.move.move.view</field>
-            <field name="model">stock.move</field>
-            <field name="inherit_id" ref="stock.view_move_form"/>
-            <field name="priority" eval="16"/>
-            <field name="arch" type="xml">
-                <data>
-                    <field name="priority" position="after">
-                        <field name="auto_move"/>
-                    </field>
-                </data>
-            </field>
-        </record>
-
-        <record model="ir.ui.view" id="stock_auto_move_move_form_picking_view">
-            <field name="name">stock.auto.move.move.form.picking.view</field>
-            <field name="model">stock.move</field>
-            <field name="inherit_id" ref="stock.view_move_picking_form"/>
-            <field name="priority" eval="16"/>
-            <field name="arch" type="xml">
-                <data>
-                    <field name="date_expected" position="after">
-                        <field name="auto_move"/>
-                    </field>
-                </data>
-            </field>
-        </record>
-
-        <record model="ir.ui.view" id="stock_auto_move_procurement_rule_form_view">
-            <field name="name">stock.auto.move.procurement.rule.form.view</field>
-            <field name="model">procurement.rule</field>
-            <field name="inherit_id" ref="stock.view_procurement_rule_form_stock_inherit"/>
-            <field name="priority" eval="16"/>
-            <field name="arch" type="xml">
-                <data>
-                    <field name="procure_method" position="after">
-                        <field name="auto_move"/>
-                    </field>
-                </data>
-            </field>
-        </record>
-
-
-    </data>
-</openerp>
