+++ stock_auto_move/views/stock_move.xml
@@ -1,0 +1,29 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<odoo>
+
+    <record model="ir.ui.view" id="stock_auto_move_move_form_view">
+        <field name="name">stock.auto.move.move.view</field>
+        <field name="model">stock.move</field>
+        <field name="inherit_id" ref="stock.view_move_form"/>
+        <field name="priority" eval="16"/>
+        <field name="arch" type="xml">
+            <field name="priority" position="after">
+                <field name="auto_move"/>
+            </field>
+        </field>
+    </record>
+
+    <record model="ir.ui.view" id="stock_auto_move_move_form_picking_view">
+        <field name="name">stock.auto.move.move.form.picking.view</field>
+        <field name="model">stock.move</field>
+        <field name="inherit_id" ref="stock.view_move_picking_form"/>
+        <field name="priority" eval="16"/>
+        <field name="arch" type="xml">
+            <field name="date_expected" position="after">
+                <field name="auto_move"/>
+            </field>
+        </field>
+    </record>
+
+</odoo>
